Split WalkPlanner class
